requirements.txt
websocket-client==1.6.4
psutil==5.9.5
requests==2.31.0
configparser==5.3.0
cryptography==41.0.7
PyQt5==5.15.10
Pillow>=10.2.0